Callbacks as third parameter for whenNext (#24)
* Callbacks for third parameter for whenNext

whenNext() and whenNextComplete() may now receive a callback to calculate the new value of a dependent cell. The implementation is not sophisticated and needs to be checked, for example if updates are still atomic?

* Callback for whenNext receives new value of `other`

The callback receiveed `this`, but a reference to this cell is available
anyway. Now `v` is passed to the callback, similar to `pred` callback.