Redefine whenNext
(1) Remove whenComplete in favour of a boolean parameter to the valueCallback to indicate, if the new value is value. The same parameter is used for onNext callbacks.
(2) Remove `pred` from whenNext. Instead, the valueCallback returns an instance of WhenNextOutcome[V], i.e. the new value for a cell and whether this value is final.