Make sure quiescentResolve-methods wait for complete quiescent
Currently, if resolve or fallback trigger new computations due to dependencies, the success of the returned Future of quiescentResolve* methods don't indicate quiescence. It only indicated that quiescence was once reached and the cycles were resolved.

With this change, those methods recursively call themselves if cycles were resolved and finialize the Future only if true quiescence is reached.